fix(video): cast pic flag constants to u32 for Windows compatibility
On Windows, bindgen generates enum constants as i32 (C default for
enum types) rather than c_uint. Add explicit cast to u32 for the
NV_ENC_PIC_FLAG_FORCEIDR and NV_ENC_PIC_FLAG_EOS re-exports.

+++ crates/rayplay-video/src/nvenc_sys.rs
@@ -138,8 +138,8 @@ pub(crate) mod ffi {
 
     // ── Pic flag re-exports ──
 
-    pub const NV_ENC_PIC_FLAG_FORCEIDR: u32 = _NV_ENC_PIC_FLAGS_NV_ENC_PIC_FLAG_FORCEIDR;
-    pub const NV_ENC_PIC_FLAG_EOS: u32 = _NV_ENC_PIC_FLAGS_NV_ENC_PIC_FLAG_EOS;
+    pub const NV_ENC_PIC_FLAG_FORCEIDR: u32 = _NV_ENC_PIC_FLAGS_NV_ENC_PIC_FLAG_FORCEIDR as u32;
+    pub const NV_ENC_PIC_FLAG_EOS: u32 = _NV_ENC_PIC_FLAGS_NV_ENC_PIC_FLAG_EOS as u32;
 
     // ── Buffer usage re-exports ──
